new solo node

+++ src/capybarauno_twist2ticks.cpp
@@ -37,8 +37,8 @@ void cmdvelCallback(const geometry_msgs::Twist::ConstPtr& twist)
         ROS_INFO("LINEAR %f ANGULAR %f ANGULAR AFTER BASELINE %f", twist->linear.x, twist->angular.z,rotational_velocity);
     }
     capybarauno::capybara_ticks ct;
-    ct.leftEncoder=(-translational_velocity+rotational_velocity)/left_meter_to_ticks;
-    ct.rightEncoder=(translational_velocity+rotational_velocity)/right_meter_to_ticks;
+    ct.leftEncoder=(uint16_t)((-translational_velocity+rotational_velocity)/left_meter_to_ticks);
+    ct.rightEncoder=(uint16_t)((translational_velocity+rotational_velocity)/right_meter_to_ticks);
     if(c.debug){
         ROS_INFO("TICKS %d %d",ct.leftEncoder, ct.rightEncoder);
     }
